use Suggests: xscreensaver for -data and -gl packages instead of Enhances.
So that people installing -data or -gl packages on their own can get the
idea that they might want to install the xscreensaver package to fully
enjoy these packages.

The -data-extra and -gl-extra packages Depend on -data and -gl
respectively, so they don't need more upwards dependencies. They Enhance
xscreensaver, so that when installing xscreensaver a sophisticated
package manager can mention them.

Since xscreensaver and gnome-screensaver* Depend, Recommend or Suggest
-data and -gl, the two latter do not need any Enhances towards the two
former.

This is the (near-future*) status of things from the point of view of
yours truly committer:

xscreensaver  gnome-screensaver   kscreensaver-xsavers
  |     |      /        \         /    |          |
 dep    |   *rec  --dep-----------    dep         |
  |     |    /   /        \            |          |
  v     |   v   v        *sug          v          |
xscreensaver-data           \---> xscreensaver-gl |
  ^     |                         ^     ^         |
  |	 \                       /      |        dep
 dep      ---------------sug----/      dep        |
  |                                     |         v
xscreensaver-data-extra        xscreensaver-gl-extra

And that's not even complete. Who said this was messy?

*) gnome-screensaver should really Recommend xscreensaver-data and
Suggest xscreensaver-data-gl, but doesn't do it yet. Currently it only
Recommends rss-glx which has nice-looking screensavers that burn all
your CPU.

Note to the (hopefully not so distant) future: When 3D becomes reliable
(or used for more than screensavers so we don't get the blame for all
driver lock-ups) on Debian, we might even have xscreensaver depend on
xscreensaver-gl. Or earlier, if xscreensaver gets an easy way to toggle
all 3D hacks on and off (maybe default to off). It's easy to say "fix
your X", but we try to be user-friendly.